pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-inbox_env"
version = "0.1.0"
description = "Inbox Env environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    # install from github
    # "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "sentence-transformers>=3.0.0",
    "scikit-learn>=1.5.0",
    "openai>=1.50.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m inbox_env.server.app
server = "inbox_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["inbox_env", "inbox_env.server"]
package-dir = { "inbox_env" = ".", "inbox_env.server" = "server" }